* moved all EDNS(0) stuff into separate module
