+++ tdns/edns0/edns0_ede.go
@@ -1,7 +1,7 @@
 /*
- * Copyright (c) 2024 Johan Stenstam, johani@johani.org
+ * Copyright (c) 2024 Johan Stenstam, johan.stenstam@internetstiftelsen.se
  */
-package tdns
+package edns0
 
 import (
 	"fmt"
