+++ tdns/edns0/go.mod
@@ -1,0 +1,13 @@
+module github.com/johanix/tdns/tdns/edns0
+
+go 1.23.2
+
+require github.com/miekg/dns v1.1.68
+
+require (
+	golang.org/x/mod v0.24.0 // indirect
+	golang.org/x/net v0.40.0 // indirect
+	golang.org/x/sync v0.14.0 // indirect
+	golang.org/x/sys v0.33.0 // indirect
+	golang.org/x/tools v0.33.0 // indirect
+)
